More wild formalizations.
Still far from the goal.  Ed reminds me that I need to relativize many
some notions to a point.  The definition of object theory structures
also seems to be wrong.  Being a concept just means exemplifying a
certain property, whereas I have taken "concept" as a primitive.
Based on my experience with Ed and Paul, "concept" is actually more
akin to a constant symbol, that is, a distinguished object.